refactor(studio): 分离观测收件箱时间线呈现
