+++ test/architecture/import-boundaries.test.ts
@@ -163,6 +163,8 @@ const RULES: ForbiddenRule[] = [
       'studio/presentation/observation-inbox/helpers.ts::observability/feedback-projection.ts',
       'studio/presentation/observation-inbox/reviewer-report.ts::observability/inbox-view-model.ts',
       'studio/presentation/observation-inbox/reviewer-report.ts::observability/feedback-projection.ts',
+      'studio/presentation/observation-inbox/timeline.ts::observability/inbox-view-model.ts',
+      'studio/presentation/observation-inbox/timeline.ts::observability/feedback-projection.ts',
       'studio/presentation/skill-health-renderer.ts::observability/skill-health-analyzer.ts',
     ],
   },
@@ -562,6 +564,22 @@ describe('架构边界守门', () => {
     expect(extractSpecifiers(renderer)).toContain('./observation-inbox/reviewer-report.js');
     expect(renderer).not.toContain('const renderReviewerReport =');
     expect(extractSpecifiers(reviewerReport)).not.toContain('../observation-inbox-renderer.js');
+  });
+
+  it('Observation Inbox 时间线呈现由独立组件拥有', () => {
+    const renderer = readFileSync(
+      join(SRC_DIR, 'studio', 'presentation', 'observation-inbox-renderer.ts'),
+      'utf-8',
+    );
+    const timeline = readFileSync(
+      join(SRC_DIR, 'studio', 'presentation', 'observation-inbox', 'timeline.ts'),
+      'utf-8',
+    );
+
+    expect(extractSpecifiers(renderer)).toContain('./observation-inbox/timeline.js');
+    expect(renderer).not.toContain('const renderTimelinePair =');
+    expect(renderer).not.toContain('const renderExperienceTimeline =');
+    expect(extractSpecifiers(timeline)).not.toContain('../observation-inbox-renderer.js');
   });
 
   it('whitelist 不能腐烂:每条 whitelist 都对应一条真实存在的 import', () => {
